requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.6.0
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
decouple==0.0.7
defusedxml==0.7.1
distro==1.9.0
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.2.2
h11==0.14.0
heyoo==0.1.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.0.3
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.40.3
packaging==24.1
pycryptodome==3.18.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.7.0
python-barcode==0.14.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
sniffio==1.3.1
sqlparse==0.4.4
svg.path==6.3
tqdm==4.66.5
typing==3.7.4.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.0.7
openpyxl==3.1.2